# [Cherry-pick] Optimize elementwise_add_grad op(#32051)